backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
python-dotenv
shapely
ultralytics
opencv-python
pydantic-settings
redis


